Make models amenable to scan (#157)
* Make models amenable to scan

We replace the `for` loop in both Llama and Mixtral with an equivalent
`HomogenousSequential` layer, which can be either run a for loop or use
`torch_xla`'s scan operator. This is a clean-ish way to turn scan on/off
without cluttering the modeling code.

I also adjusted Mixtral slightly so that we can even run `scan` in
Mixtral with its static MoE implementation. Scanning over GMM on the
other hand won't work until GMM forward/backward is wrapped in a custom
op similar to https://github.com/pytorch/xla/pull/8654.

Test: added unit test. Next PR will change the trainer to apply scan.

* Address comments